feat(server): moved dynamic secret to ee
